back/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
python-decouple==3.8
celery==5.4.0
redis==5.0.6
django-storages==1.14.3
boto3==1.34.162
django-extensions==3.2.3
django-debug-toolbar==4.4.6
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.6.1
black==24.8.0
flake8==7.1.1
isort==5.13.2
psycopg[binary]==3.2.3
